feat: recent ideas block with working settings
Implements finished fetching and uses settings from the content.
Renames recent stories block to recent ideas to match the new naming.
